Make RenderExtensionMode enum class
It's odd otherwise, to have names like Count leak in to QSSGRenderLayer.

+++ src/runtimerender/rendererimpl/qssglayerrenderdata.cpp
@@ -1945,8 +1945,8 @@ void QSSGLayerRenderData::prepareForRender()
     {
         // Give user provided passes a chance to modify the renderable data before starting
         // Note: All non-active extensions should be filtered out by now
-        Q_STATIC_ASSERT(USERPASSES == QSSGRenderLayer::RenderExtensionMode::Count);
-        for (int i = 0; i != QSSGRenderLayer::RenderExtensionMode::Count; ++i) {
+        Q_STATIC_ASSERT(USERPASSES == size_t(QSSGRenderLayer::RenderExtensionMode::Count));
+        for (size_t i = 0; i != size_t(QSSGRenderLayer::RenderExtensionMode::Count); ++i) {
             const auto &renderExtensions = layer.renderExtensions[i];
             auto &userPass = userPasses[i];
             for (auto rit = renderExtensions.crbegin(), rend = renderExtensions.crend(); rit != rend; ++rit) {
@@ -2051,7 +2051,7 @@ void QSSGLayerRenderData::prepareForRender()
     if (thePrepResult.flags.requiresScreenTexture())
         activePasses.push_back(&screenMapPass);
 
-    auto &underlayPass = userPasses[QSSGRenderLayer::RenderExtensionMode::Underlay];
+    auto &underlayPass = userPasses[size_t(QSSGRenderLayer::RenderExtensionMode::Underlay)];
     if (underlayPass.hasData())
         activePasses.push_back(&underlayPass);
 
@@ -2079,7 +2079,7 @@ void QSSGLayerRenderData::prepareForRender()
     if (transparentObjects.size() > 0 || (!layerEnableDepthTest && hasOpaqueObjects))
         activePasses.push_back(&transparentPass);
 
-    auto &overlayPass = userPasses[QSSGRenderLayer::RenderExtensionMode::Overlay];
+    auto &overlayPass = userPasses[size_t(QSSGRenderLayer::RenderExtensionMode::Overlay)];
     if (overlayPass.hasData())
         activePasses.push_back(&overlayPass);
